tests: Allow running tests in parallel.
Running the tests in parallel may reduce the time to complete the
testsuite.

With the standard GitHub hosted 4 VCPU runner, there are not
enough resources available to benefit, but on a more performant
system running all 4 tests in parallel the wall clock time is
reduced from 9m 30s to 3m 52s.

Signed-off-by: Frode Nordahl <fnordahl@ubuntu.com>