Further testing of the tls integration
We now fully test that a application can interface with the microovn
units using certificates given to it from the self-signed-certificates
provider. To do this interface-consumer now is able to integrate with
self-signed-certificates, obtain the CA cert file and a certificate and
private key assigned to it. These are then used to verify end to end
connection as part of the test.

Signed-off-by: MJ Ponsonby <mj.ponsonby@canonical.com>