Modify workflows to use charmcraftcache
Adds a _build_charm workflow from the data platforms team that is
modified to include a pre-pack-hook allowing us to properly fetch
charm-libs and ensures we only pack for a single platform as that is all
we are using in our test-suites.

This also sets the tests to run in parallel which significantly speeds
up the ci.

Signed-off-by: MJ Ponsonby <mj.ponsonby@canonical.com>
(cherry picked from commit fcc8704f2b5d825662a3d3d2af637a2006c15cff)